+++ src/file/source-map/clean-source-map.js
@@ -1,4 +1,4 @@
-var getSourceFilePath = require('./get-source-file-path');
+var path = require('path');
 
 function clean(map, siteRoot) {
 
@@ -20,6 +20,29 @@ function cleanSources(sources, siteRoot) {
     return sources.map(function(source) {
         return getSourceFilePath(source, siteRoot);
     });
+
+}
+
+function getSourceFilePath(filePath, siteRoot) {
+
+    var sourceMapRoot = getSourceMapRoot(filePath, siteRoot),
+        sourceMapPath = path.join(sourceMapRoot, path.basename(filePath)).replace(/\\/g, '/');
+
+    if (!sourceMapPath.startsWith('/')) {
+        sourceMapPath = '/' + sourceMapPath;
+    }
+
+    return sourceMapPath;
+
+}
+
+function getSourceMapRoot(filePath, siteRoot) {
+
+    var directory = path.normalize(path.dirname(filePath));
+
+    siteRoot = path.normalize(siteRoot);
+
+    return directory.replace(siteRoot, '');
 
 }
